requirements.txt
flask==2.2.3
google-genai==1.11.0
python-dotenv==1.1.0
urllib3==2.0.3